Move classes to seperate project
