+++ cohort-cql-optimizations/src/main/java/com/ibm/cohort/cql/evaluator/ShortOrEvaluator.java
@@ -1,11 +1,9 @@
 /*
+ * (C) Copyright IBM Corp. 2021, 2021
  *
- *  * (C) Copyright IBM Corp. 2021
- *  *
- *  * SPDX-License-Identifier: Apache-2.0
- *
+ * SPDX-License-Identifier: Apache-2.0
  */
-package com.ibm.cohort.engine.elm.execution;
+package com.ibm.cohort.cql.evaluator;
 
 import org.cqframework.cql.elm.execution.Or;
 import org.opencds.cqf.cql.engine.elm.execution.OrEvaluator;
