+++ cohort-cql-optimizations/src/test/resources/cql/short-circuit-and-or.xml
