+++ cohort-engine/pom.xml
@@ -33,6 +33,12 @@
 		</dependency>
 
 		<dependency>
+			<groupId>com.ibm.cohort</groupId>
+			<artifactId>cohort-cql-optimizations</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
 			<groupId>info.cqframework</groupId>
 			<artifactId>elm</artifactId>
 		</dependency>
@@ -113,12 +119,6 @@
 			<artifactId>jackson-core</artifactId>
 		</dependency>
 		
-		<dependency>
-			<groupId>xpp3</groupId>
-			<artifactId>xpp3</artifactId>
-			<version>1.1.4c</version>
-		</dependency>
-
 		<!-- Runtime dependencies -->
 
 		<dependency>
