+++ cohort-evaluator/pom.xml
@@ -22,7 +22,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>com.ibm.cohort</groupId>
-			<artifactId>cohort-engine</artifactId>
+			<artifactId>cohort-cql-optimizations</artifactId>
 			<version>${project.version}</version>
 		</dependency>
